godeps: Upgrade kubernetes client libraries 1.20.4→1.21.9 [ci-skip]
Besides just a search/replace for the version number, this also involved:

 - rebasing github.com/emissary-ingress/code-generator
 - adding `option go_package` to our .proto files, since the newer
   protoc-gen-go that will get pulled in is pickier about them (thanks
   @aidanhahn for this one)

I went to 1.21 instead of 1.22 or 1.23 because
k8s.ip/kube-openapi@4e36c7e0cbd6d08c871dc44f2d2be29108874175 (which is
in the 1.22 branch) changes the meaning of "spec" in

    type OpenAPIDefinition struct {
        Schema       spec.Schema
        Dependencies []string
    }

from "github.com/go-openapi/spec" to
"k8s.io/kube-openapi/pkg/validation/spec", which breaks lots of
external users, such as the Argo libraries that the agent uses.

So stick with 1.21.x for now.

Signed-off-by: Aidan Hahn <aidanhahn@datawire.io>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>