deps: upgrade protoc to enable building on M1 mac (#4202)
Prior to this change the version of protoc being used was `3.8.0`. An arm64 version of protoc for osx was not available in `3.8.0` so running commands like `make generate` do not work for contributors running on M1 macs.

This upgrades the repository to the latest version of protoc and fixes the `tools.mk` so that it can download protoc correctly. The provided protoc binaries that are released for osx use the name schema of `osx_aarch_64` so this fix addresses this difference between what is reported by uname.

Signed-off-by: Lance Austin <laustin@datawire.io>